+++ .github/workflows/win.yml
@@ -50,7 +50,7 @@ jobs:
     - name: Build project dependencies
       run: |
         python -m pip install "torch==1.8.0+cpu" -f https://download.pytorch.org/whl/torch_stable.html
-        python -m pip install "gym>=0.17.3,<0.18.0" "stable_baselines3>=0.10" "importlib-metadata>=3.3.0"
+        python -m pip install "gym>=0.18.3" "stable_baselines3>=0.10" "importlib-metadata>=3.3.0"
         & "./build_tools/build_install_deps_windows.ps1"
 
     #####################################################################################
