+++ python/gym_jiminy/toolbox/gym_jiminy/toolbox/math/__init__.py
@@ -1,8 +1,15 @@
 # pylint: disable=missing-module-docstring
 
 from .spline import Spline
+from .qhull import ConvexHull
+from .generic import squared_norm_2
+from .signal import integrate_zoh, smoothing_filter
 
 
 __all__ = [
-    "Spline"
+    "Spline",
+    "ConvexHull",
+    "squared_norm_2",
+    "integrate_zoh",
+    "smoothing_filter"
 ]
