+++ python/gym_jiminy/toolbox/gym_jiminy/toolbox/wrappers/__init__.py
@@ -1,11 +1,11 @@
 # pylint: disable=missing-module-docstring
 
 from .normal_action import NormalizeAction
-from .meta_envs import HierarchicalTaskSettableEnv, CurriculumTaskWrapper
+from .meta_envs import HierarchicalTaskSettableEnv, TaskSchedulingWrapper
 
 
 __all__ = [
     "NormalizeAction",
     "HierarchicalTaskSettableEnv",
-    "CurriculumTaskWrapper"
+    "TaskSchedulingWrapper"
 ]
