+++ python/jiminy_py/setup.py
@@ -94,6 +94,7 @@ setup(
         # Standard library to generate figures.
         f"matplotlib{matplotlib_spec}",
         # Used internally to read HDF5 format log files.
+        # This module is NOT supported by PyPy.
         "h5py",
         # Used internally by Robot to replace meshes by associated minimal
         # volume bounding box.
@@ -106,7 +107,9 @@ setup(
         # collision bodies.
         "meshcat>=0.0.19",
         # Standalone mesh visualizer used as Viewer's backend.
-        # 1.10.9 adds support of Nvidia EGL rendering without X11 server.
+        # Panda3d>1.10.9 adds support of Nvidia EGL rendering without X11
+        # server.
+        # Panda3d is NOT supported by PyPy.
         "panda3d_viewer",
         # Used internally by Viewer to record video programmatically when
         # Meshcat is not used as rendering backend.
